+++ src/Command/Redis/ZMSCORE.php
@@ -12,7 +12,7 @@
 
 namespace Predis\Command\Redis;
 
-use Predis\Command\Command as RedisCommand;
+use Predis\Command\PrefixableCommand as RedisCommand;
 
 /**
  * @see https://redis.io/commands/zmscore/
@@ -30,5 +30,10 @@ class ZMSCORE extends RedisCommand
     public function getId()
     {
         return 'ZMSCORE';
+    }
+
+    public function prefixKeys($prefix)
+    {
+        $this->applyPrefixForFirstArgument($prefix);
     }
 }