+++ tests/Predis/Cluster/RedisStrategyTest.php
@@ -503,6 +503,7 @@ class RedisStrategyTest extends PredisTestCase
             'ZREVRANGEBYSCORE' => 'keys-first',
             'ZREVRANK' => 'keys-first',
             'ZSCORE' => 'keys-first',
+            'ZMSCORE' => 'keys-first',
             'ZUNIONSTORE' => 'keys-zaggregated',
             'ZSCAN' => 'keys-first',
             'ZLEXCOUNT' => 'keys-first',